+++ heudiconv/dicoms.py
@@ -89,7 +89,8 @@ def create_seqinfo(mw, series_files, series_id):
         patient_age=dcminfo.get('PatientAge'),
         patient_sex=dcminfo.get('PatientSex'),
         date=dcminfo.get('AcquisitionDate'),
-        series_uid=dcminfo.get('SeriesInstanceUID')
+        series_uid=dcminfo.get('SeriesInstanceUID'),
+        time=dcminfo.get('AcquisitionTime'),
     )
     return seqinfo
 
@@ -265,8 +266,13 @@ def group_dicoms_into_seqinfos(files, grouping, file_filter=None,
         series_id = '-'.join(map(str, series_id))
         if mw.image_shape is None:
             # this whole thing has no image data (maybe just PSg DICOMs)
-            # nothing to see here, just move on
-            continue
+            # If this is a Siemens PhoenixZipReport or PhysioLog, keep it:
+            if mw.dcm_data.SeriesDescription == 'PhoenixZIPReport':
+                # give it a dummy shape, so that we can continue:
+                mw.image_shape = (0, 0, 0)
+            else:
+                # nothing to see here, just move on
+                continue
         seqinfo = create_seqinfo(mw, series_files, series_id)
 
         if per_studyUID: