+++ heudiconv/tests/test_heuristics.py
@@ -7,8 +7,13 @@ from six.moves import StringIO
 
 from glob import glob
 from os.path import join as pjoin, dirname
+from pathlib import Path
 import csv
 import re
+
+from .. import __version__
+from ..bids import HEUDICONV_VERSION_JSON_KEY
+from ..utils import load_json
 
 import pytest
 from .utils import TESTS_DATA_PATH
@@ -140,16 +145,22 @@ def test_scout_conversion(tmpdir):
     ).split(' ') + ['-o', tmppath]
     runner(args)
 
-    assert not op.exists(pjoin(
-        tmppath,
-        'Halchenko/Yarik/950_bids_test4/sub-phantom1sid1/ses-localizer/anat'))
+    dspath = Path(tmppath) / 'Halchenko/Yarik/950_bids_test4'
+    sespath = dspath / 'sub-phantom1sid1/ses-localizer'
 
-    assert op.exists(pjoin(
-        tmppath,
-        'Halchenko/Yarik/950_bids_test4/sourcedata/sub-phantom1sid1/'
-        'ses-localizer/anat/sub-phantom1sid1_ses-localizer_scout.dicom.tgz'
-    )
-    )
+    assert not (sespath / 'anat').exists()
+    assert (
+        dspath /
+        'sourcedata/sub-phantom1sid1/ses-localizer/'
+        'anat/sub-phantom1sid1_ses-localizer_scout.dicom.tgz'
+    ).exists()
+
+    # Let's do some basic checks on produced files
+    j = load_json(sespath / 'fmap/sub-phantom1sid1_ses-localizer_acq-3mm_phasediff.json')
+    # We store HeuDiConv version in each produced .json file
+    # TODO: test that we are not somehow overwritting that version in existing
+    # files which we have not produced in a particular run.
+    assert j[HEUDICONV_VERSION_JSON_KEY] == __version__
 
 
 @pytest.mark.parametrize(
@@ -176,3 +187,20 @@ def test_notop(tmpdir, bidsoptions):
             assert not op.exists(pjoin(tmppath, 'Halchenko/Yarik/950_bids_test4', fname))
         else:
             assert op.exists(pjoin(tmppath, 'Halchenko/Yarik/950_bids_test4', fname))
+
+
+def test_phoenix_doc_conversion(tmpdir):
+    tmppath = tmpdir.strpath
+    subID = 'Phoenix'
+    args = (
+        "-c dcm2niix -o %s -b -f bids_PhoenixReport --files %s -s %s"
+        % (tmpdir, pjoin(TESTS_DATA_PATH, 'Phoenix'), subID)
+    ).split(' ')
+    runner(args)
+
+    # check that the Phoenix document has been extracted (as gzipped dicom) in
+    # the sourcedata/misc folder:
+    assert op.exists(pjoin(tmppath, 'sourcedata', 'sub-%s', 'misc', 'sub-%s_phoenix.dicom.tgz') % (subID, subID))
+    # check that no "sub-<subID>/misc" folder has been created in the BIDS
+    # structure:
+    assert not op.exists(pjoin(tmppath, 'sub-%s', 'misc') % subID)